Harden Publish skript

+++ publish.ps1
@@ -34,7 +34,11 @@ New-Item -ItemType directory -Path $PathToNewFolder
 Copy-Item .\* $PathToNewFolder -Recurse
 
 Remove-Item $PathToNewFolder\publish.ps1
-Remove-Item $PathToNewFolder\.git -Recurse -Force
+
+if (Test-Path $PathToNewFolder\.git)
+{
+  Remove-Item $PathToNewFolder\.git -Recurse -Force
+}
 
 $ArchivePath = "..\Versendete_Versionen\$($NameOfNewFolder).zip"
 Compress-Archive -Path $PathToNewFolder -DestinationPath $ArchivePath
